evm-jira/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.5
    # via google-auth
defusedxml==0.7.1
    # via jira
google-auth==2.48.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.4
    # via gspread
gspread==6.2.1
    # via -r requirements.in
idna==3.11
    # via requests
jira==3.10.5
    # via -r requirements.in
numpy==2.4.2
    # via pandas
oauthlib==3.3.1
    # via requests-oauthlib
packaging==26.0
    # via jira
pandas==3.0.0
    # via -r requirements.in
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
requests==2.32.5
    # via
    #   jira
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via
    #   google-auth-oauthlib
    #   jira
requests-toolbelt==1.0.0
    # via jira
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
typing-extensions==4.15.0
    # via jira
urllib3==2.6.3
    # via requests
